requirements.txt
Flask==2.1.2
clearstreet.logger==2.0.2
boto3==1.24.47
botocore==1.27.48
gestalt-cfg==3.0.0
